requirements.txt
bson==0.5.10
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.3.2
python-dateutil==2.8.2
python-dotenv==0.21.0
six==1.16.0
Werkzeug==2.2.2
